1.2.21 - Working version with two form appearance modes

+++ WeatherStation/PreferencesForm.cs
@@ -126,8 +126,8 @@ namespace WeatherStation
 
 
                 //Store to vars main interface settings
-                ParentMainForm.maxNumberOfPointsInChart = Convert.ToInt16(txtMaxPoints.Text);
-                ParentMainForm.timer_main.Interval = Convert.ToInt16(txtRefreshInterval.Text);
+                ParentMainForm.maxNumberOfPointsInChart = Convert.ToInt32(txtMaxPoints.Text);
+                ParentMainForm.timer_main.Interval = Convert.ToInt32(txtRefreshInterval.Text);
                 ParentMainForm.timer_debug_changetext.Interval = Convert.ToInt16(txtRefreshInterval.Text);
 
                 //Store webservice settings
